chore(fe): move datatable to main
