+++ apps/frontend/app/(client)/(main)/contest/[contestId]/@tabs/announcement/page.tsx
@@ -1,4 +1,4 @@
-import DataTable from '@/components/DataTable'
+import DataTable from '@/app/(client)/(main)/_components/DataTable'
 import { fetcher } from '@/lib/utils'
 import type { ContestAnnouncement } from '@/types/type'
 import { columns } from './_components/Columns'
