+++ apps/frontend/app/(client)/(main)/contest/_components/FinishedContestTable.tsx
@@ -1,4 +1,4 @@
-import DataTable from '@/components/DataTable'
+import DataTable from '@/app/(client)/(main)/_components/DataTable'
 import { fetcher, fetcherWithAuth } from '@/lib/utils'
 import type { Contest } from '@/types/type'
 import type { Session } from 'next-auth'
